+++ Modules/Ruleset.IP/Test/Public/ConvertTo-Subnet.test.ps1
@@ -97,7 +97,6 @@ InModuleScope Ruleset.IP {
 		# Default mocks
 		#
 
-
 		#
 		# Tests
 		#